Relative Links header for retrieving collections
Has links for 'first', 'prev', 'next', and 'last'.
Only added when it's a meaningful navigation,
URIs for same page or beyond valid range is omitted.

Uses route with most variable matches and variables are not repeated as
query parameters.  e.g. if it contains "/:page/" it will be chosen over
a shorter route plus "?page=" query.

URIs are relative to site's base URL.
It is for the client to append them appropriately.
This is necessary because of caching concerns, the API needs to be
impartial to the domain name and always respond the same.