Set the headers `Cache-Control`, `Last-Modified`, and `Vary`
`Cache-Control` is only public when request is unauthenticated.
`Last-Modified` is only shown for entities.
Collections can change so that the Last-Modified goes backwards so let's
avoid that use case.
`Vary` needs to be the same for each resource because there is no way to
know who will request it next.
Thus it is always "Accept,Authorization,Version".
"Accept" to allow for XML/JSON.
"Authorization" to allow for user types.
"Version" because you never know when a new version will come.