URL rewrites' `id_path` and `description` are not visible
`description` is not used.
`id_path` does not need to be published.
It's only use is to build an URL for categories and products so can be
deduced from `category_id` and `product_id`.
For other rewrites it is effectively random.
It could be conceivably be dropped from core entirely.