Product options with 'select' types must have `values` field
This was previously handled as a core exception after saving the option.
Now it is part of validation before saving.
That allows multiple error messages to be returned simultaneously.
If the product option is not required and `values` is absent then an
empty array is assumed and no error occurs.